chore: refactor controlplane client
