+++ agent/common/controlplane/controlplane.go
@@ -22,7 +22,7 @@ import (
 	"io"
 	"net/http"
 
-	"github.com/eclipse-cfm/cfm/agent/orchestration/edcv"
+	vault "github.com/eclipse-cfm/cfm/agent/common/vault"
 	"github.com/eclipse-cfm/cfm/common/token"
 )
 
@@ -44,7 +44,7 @@ type ParticipantContextConfig struct {
 	SecretEntries        map[string]string `json:"privateEntries"`
 }
 
-func NewParticipantContextConfig(participantContextID string, stsClientID string, stsClientSecretAlias string, participantID string, vConfig edcv.VaultConfig, stsTokenURL string) ParticipantContextConfig {
+func NewParticipantContextConfig(participantContextID string, stsClientID string, stsClientSecretAlias string, participantID string, vConfig vault.Config, stsTokenURL string) ParticipantContextConfig {
 	vaultConfig := map[string]any{
 		"config": vConfig,
 	}
