+++ agent/orchestration/edcv/activity/activity_test.go
@@ -17,7 +17,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/eclipse-cfm/cfm/agent/orchestration/edcv/controlplane"
+	"github.com/eclipse-cfm/cfm/agent/common/controlplane"
 	"github.com/eclipse-cfm/cfm/common/model"
 	"github.com/eclipse-cfm/cfm/common/system"
 	"github.com/eclipse-cfm/cfm/pmanager/api"
